Make set_value protected
Provide three public interfaces:
- value
- dvalue
- value_and_dvalue

`value` and `value_and_dvalue` are both prefered as `Model`s always only override `set_value` which
evaluate BOTH the forward operator and its derivative. Calling `dvalue` alone will waste the model output.

I was trying to think of a better as mentioned in #5, but I failed. The real issue is that I can't come up
with a name set that is consistent for all these three public interfaces. For example
- forward
- D
- forward_and_D

looks weird. But I am fine with
- forward
- dforward
- forward_and_dforward

I am okay with this if you think it plays better with torch idioms.

Possibly close #5, we'll see.