Add an interface to take care of side effects of advancing time step
Currently, I let the predictor shift the states back in time after every solve. This is fine as long as every "time step" has exactly one "constitutive update".

However, in the context of finite element method and optimization there might be an outer loop, and it is preferrable to let the driver program to tell NEML2 models when to advance the time step.