Implement MOOSE-like viscoplasticity models
If one assumes associative flow and J2 plasticity, the return mapping can be significantly simplified: the implicit system always only has 1 dof for the consistency parameter. This is because the flow direction does not change during the return mapping. To achieve this, we need several new objects (implemented in this PR) and a modified way of composing the models (see the new regression tests).

close #101